# docs/starlight/public/openapi/gateway.yaml
openapi: 3.1.0
info:
  title: ProtocolSoup Gateway API
  version: 1.0.0
  summary: API gateway and protocol aggregation service.
  description: |
    Public contract for the ProtocolSoup gateway service.
    The gateway aggregates protocol metadata and proxies protocol-specific traffic to upstream services.
servers:
  - url: https://protocolsoup.com
    description: Production gateway
security: []
...
paths:
...
  /api/protocols:
    get:
      tags: [Discovery]
      summary: List aggregated protocols
      operationId: listGatewayProtocols
      responses:
        "200":
          description: Aggregated protocol list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtocolListResponse"
        "503":
          description: No upstream protocols available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/protocols/{id}/flows:
    get:
      tags: [Discovery]
      summary: List protocol flow definitions
      operationId: getGatewayProtocolFlows
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Flow definitions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowListResponse"
        "404":
          description: Protocol not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/protocols/{id}/demo/{flow}:
    post:
      tags: [LookingGlass]
      summary: Start a protocol demo session
      operationId: startGatewayDemo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: flow
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartDemoResponse"
              examples:
                demo:
                  value:
                    session_id: "session-abc123"
                    session_token: "JmRf8mW0gMmT_5kXc4mBAc0f7Zt5F8yQ4xVvYxjv8Eg"
                    protocol: "scim"
                    flow: "user-lifecycle"
                    ws_endpoint: "/ws/lookingglass/session-abc123"
                    scenario: null
        "404":
          description: Protocol or flow not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/lookingglass/sessions:
    get:
      tags: [LookingGlass]
      summary: List Looking Glass sessions
      description: Returns metadata only; captured events are never included in this response.
      operationId: listGatewayLookingGlassSessions
      responses:
        "200":
          description: Session list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/LookingGlassSessionSummary"
        "503":
          description: Looking Glass unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/palette/query:
    post:
      tags: [Palette]
      summary: Query the content palette index
      description: >-
        Backs the homepage search bar and the global cmd+K palette. Runs a
        ranked retrieval over the prebuilt `palette.db` content index
        (protocols, flows, concepts, spec assertions) and returns results
        with match reasons, axis chips, and refinement chips. Every result
        carries its full markdown body so the selected row can render inline
        without a second round-trip; the backend never ships HTML. The route
        is only mounted when a palette index is configured.
      operationId: queryGatewayPalette
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaletteQueryRequest"
            examples:
              free_text:
                value:
                  q: "refresh token rotation"
                  limit: 10
              scoped:
                value:
                  q: "device"
                  scope: flow
                  filters:
                    - axis: protocol
                      value: oauth2
      responses:
        "200":
          description: Ranked palette results.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaletteQueryResponse"
        "400":
          description: Request body is not valid JSON.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          description: Method other than POST.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Query execution failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ws/lookingglass/{session}:
    get:
      tags: [LookingGlass]
      summary: Subscribe to the owner-visible Looking Glass event stream
      description: >-
        Upgrades to WebSocket. Clients must offer both
        `protocolsoup-lookingglass-v1` and
        `protocolsoup-lookingglass-owner.<session_token>` in
        `Sec-WebSocket-Protocol`; the owner capability is not accepted in the
        URL query string. The server selects `protocolsoup-lookingglass-v1`.
      operationId: subscribeGatewayLookingGlassSession
      parameters:
        - in: path
          name: session
          required: true
          schema:
            type: string
        - in: header
          name: Sec-WebSocket-Protocol
          required: true
          schema:
            type: string
          example: "protocolsoup-lookingglass-v1, protocolsoup-lookingglass-owner.JmRf8mW0gMmT_5kXc4mBAc0f7Zt5F8yQ4xVvYxjv8Eg"
      responses:
        "101":
          description: WebSocket upgrade accepted.
        "401":
          description: Missing or invalid owner subprotocol capability.
components:
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ProtocolListResponse:
      type: object
      required: [protocols]
      properties:
        protocols:
          type: array
          items:
            $ref: "#/components/schemas/ProtocolSummary"
    ProtocolSummary:
      type: object
      required: [id, name, version, description, tags]
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    FlowListResponse:
      type: object
      required: [flows]
      properties:
        flows:
          type: array
          items:
            type: object
            additionalProperties: true
    StartDemoResponse:
      type: object
      required: [session_id, session_token, protocol, flow, ws_endpoint]
      properties:
        session_id:
          type: string
        session_token:
          type: string
          minLength: 43
          maxLength: 43
          description: >-
            256-bit owner capability returned only when the session is created.
            Keep it confidential and use it for owner-only session APIs.
          x-sensitive: true
        protocol:
          type: string
        flow:
          type: string
        ws_endpoint:
          type: string
        scenario:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
    LookingGlassSessionSummary:
      type: object
      required: [id, protocol_id, flow_id, state, created_at, updated_at]
      properties:
        id:
          type: string
        protocol_id:
          type: string
        flow_id:
          type: string
        state:
          type: string
          enum: [active, paused, complete]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PaletteQueryRequest:
      type: object
      required: [q]
      properties:
        q:
          type: string
          description: Free-text query. Axis values and aliases inside the text are resolved server-side.
        scope:
          type: string
          enum: [flow, protocol, spec, concept]
          description: Restrict results to one artefact type. Omit for no scope filter.
        filters:
          type: array
          description: Additional (axis, value) constraints; AND-ed together.
          items:
            $ref: "#/components/schemas/PaletteAxisValue"
        limit:
          type: integer
          minimum: 1
          maximum: 50
          default: 20
          description: Values outside 1..50 fall back to the default of 20.
    PaletteAxisValue:
      type: object
      required: [axis, value]
      properties:
        axis:
          type: string
        value:
          type: string
    PaletteQueryResponse:
      type: object
      required: [query, results, refinement_chips, resolved_aliases, total_candidates, elapsed_micros]
      properties:
        query:
          type: string
          description: The original query text as received.
        results:
          type: array
          items:
            $ref: "#/components/schemas/PaletteResult"
        refinement_chips:
          type: array
          description: Suggested axis filters shown when intent is ambiguous along an axis.
          items:
            $ref: "#/components/schemas/PaletteRefinementChip"
        resolved_aliases:
          type: array
          description: Query tokens that matched a known alias and how they were resolved.
          items:
            $ref: "#/components/schemas/PaletteResolvedAlias"
        total_candidates:
          type: integer
          minimum: 0
          description: Number of candidates considered before ranking and truncation.
        elapsed_micros:
          type: integer
          minimum: 0
    PaletteResult:
      type: object
      required: [id, type, name, axis_chips, match_reasons, runnable, href, score]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [protocol, flow, concept, walkthrough, spec-assertion]
        name:
          type: string
        summary:
          type: string
          description: Frontmatter `summary`, written as a navigation label.
        protocol:
          type: string
        axis_chips:
          type: array
          items:
            $ref: "#/components/schemas/PaletteAxisValue"
        match_reasons:
          type: array
          items:
            $ref: "#/components/schemas/PaletteMatchReason"
        runnable:
          type: boolean
        href:
          type: string
          description: Canonical page for protocol and flow artefacts. Empty for inline-only types (concept, walkthrough, spec-assertion).
        run_url:
          type: string
          description: Looking Glass deep-link for runnable flows.
        score:
          type: number
        status:
          type: string
        body_preview:
          type: string
          description: Plain-text first-paragraph excerpt of the body.
        body:
          type: string
          description: Full markdown body of the artefact. Never contains HTML.
        snippet:
          type: string
          description: Query-aware excerpt of `body` with matching tokens wrapped in `<mark>` spans. Empty when no free token hit the body.
        normative_anchors:
          type: array
          items:
            $ref: "#/components/schemas/PaletteNormativeAnchor"
        related_concepts:
          type: array
          description: Artefact ids of related concepts.
          items:
            type: string
        spec_assertion:
          $ref: "#/components/schemas/PaletteSpecAssertion"
    PaletteMatchReason:
      type: object
      required: [kind, weight]
      properties:
        kind:
          type: string
          enum: [axis, alias, fts, edge, runnable, protocol-name]
        axis:
          type: string
        value:
          type: string
        artefact:
          type: string
        matched_token:
          type: string
        matched_phrase:
          type: string
        edge_kind:
          type: string
        weight:
          type: number
    PaletteRefinementChip:
      type: object
      required: [axis, value, count]
      properties:
        axis:
          type: string
        value:
          type: string
        count:
          type: integer
          minimum: 0
    PaletteResolvedAlias:
      type: object
      required: [matched_token]
      properties:
        matched_token:
          type: string
        axis:
          type: string
        value:
          type: string
        artefact:
          type: string
    PaletteNormativeAnchor:
      type: object
      required: [rfc, sections]
      properties:
        rfc:
          type: string
        sections:
          type: array
          items:
            type: string
    PaletteSpecAssertion:
      type: object
      required: [normative_level, assertion_text, anchors]
      properties:
        normative_level:
          type: string
          enum: [MUST, SHOULD, MAY, MUST NOT, SHOULD NOT]
        assertion_text:
          type: string
        anchors:
          type: array
          items:
            $ref: "#/components/schemas/PaletteNormativeAnchor"